requirements.txt
typer[all]
pydantic
pydantic-settings
pyyaml
torch
torchvision
albumentations
tqdm
pillow
numpy
